# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_Zapbox-API_evolution-go_pkg_chat_service.BodyStruct:
    properties:
      chat:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_community_service.AddParticipantStruct:
    properties:
      communityJid:
        type: string
      groupJid:
        items:
          type: string
        type: array
    type: object
  github_com_Zapbox-API_evolution-go_pkg_community_service.CreateCommunityStruct:
    properties:
      communityName:
        type: string
    type: object
...
  github_com_Zapbox-API_evolution-go_pkg_group_service.CreateGroupStruct:
    properties:
      groupName:
        type: string
      participants:
        items:
          type: string
        type: array
    type: object
  github_com_Zapbox-API_evolution-go_pkg_group_service.GetGroupInfoStruct:
    properties:
      groupJid:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_group_service.GetGroupInviteLinkStruct:
    properties:
      groupJid:
        type: string
      reset:
        type: boolean
    type: object
  github_com_Zapbox-API_evolution-go_pkg_group_service.JoinGroupStruct:
    properties:
      code:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_group_service.SetGroupNameStruct:
    properties:
      groupJid:
        type: string
      name:
        type: string
    type: object
  github_com_Zapbox-API_evolution-go_pkg_group_service.SetGroupPhotoStruct:
    properties:
      groupJid:
        type: string
      image:
        type: string
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
  /chat/archive:
    post:
      consumes:
      - application/json
      description: Archive a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Archive a chat
      tags:
      - Chat
  /chat/mute:
    post:
      consumes:
      - application/json
      description: Mute a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Mute a chat
      tags:
      - Chat
  /chat/pin:
    post:
      consumes:
      - application/json
      description: Pin a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Pin a chat
      tags:
      - Chat
  /chat/unpin:
    post:
      consumes:
      - application/json
      description: Unpin a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Unpin a chat
      tags:
      - Chat
  /community/add:
    post:
      consumes:
      - application/json
      description: Add participant to community
      parameters:
      - description: Participant data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_community_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Add participant to community
      tags:
      - Community
  /community/create:
    post:
      consumes:
      - application/json
      description: Create community
      parameters:
      - description: Community data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_community_service.CreateCommunityStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create community
      tags:
      - Community
  /community/remove:
    post:
      consumes:
      - application/json
      description: Remove participant from community
      parameters:
      - description: Participant data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_community_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Remove participant from community
      tags:
      - Community
  /group/create:
    post:
      consumes:
      - application/json
      description: Create group
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_group_service.CreateGroupStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create group
      tags:
      - Group
  /group/info:
    post:
      consumes:
      - application/json
      description: Get group info
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_group_service.GetGroupInfoStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get group info
      tags:
      - Group
  /group/invitelink:
    post:
      consumes:
      - application/json
      description: Get group invite link
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_group_service.GetGroupInviteLinkStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get group invite link
      tags:
      - Group
  /group/join:
    post:
      consumes:
      - application/json
      description: Join group link
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_group_service.JoinGroupStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Join group link
      tags:
      - Group
  /group/list:
    get:
      consumes:
      - application/json
      description: List groups
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List groups
      tags:
      - Group
  /group/myall:
    get:
      consumes:
      - application/json
      description: Get my groups
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get my groups
      tags:
      - Group
  /group/name:
    post:
      consumes:
      - application/json
      description: Set group name
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_group_service.SetGroupNameStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set group name
      tags:
      - Group
  /group/participant:
    post:
      consumes:
      - application/json
      description: Update participant
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_group_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update participant
      tags:
      - Group
  /group/photo:
    post:
      consumes:
      - application/json
      description: Set group photo
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_group_service.SetGroupPhotoStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set group photo
      tags:
      - Group
  /instance/all:
    get:
      consumes:
      - application/json
      description: Get all instances
      produces:
      - application/json
      responses:
        "200":
          description: All instances
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all instances
      tags:
      - Instance
  /instance/connect:
    post:
      consumes:
      - application/json
      description: Connect to instance with the provided data
      parameters:
      - description: Instance data
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_instance_service.ConnectStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Instance connected successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Connect to instance
      tags:
      - Instance
  /instance/create:
    post:
      consumes:
      - application/json
      description: Creates a new instance with the provided data
      parameters:
      - description: Instance data
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_instance_service.CreateStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Instance created successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new instance
      tags:
      - Instance
  /instance/delete/{instanceId}:
    delete:
      consumes:
      - application/json
      description: Delete instance
      parameters:
      - description: Instance Id
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Instance deleted successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete instance
      tags:
      - Instance
  /instance/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnect from instance
      produces:
      - application/json
      responses:
        "200":
          description: Instance disconnected successfully
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Disconnect from instance
      tags:
      - Instance
  /instance/logout:
    delete:
      consumes:
      - application/json
      description: Logout from instance
      produces:
      - application/json
      responses:
        "200":
          description: Instance logged out successfully
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Logout from instance
      tags:
      - Instance
  /instance/pair:
    post:
      consumes:
      - application/json
      description: Request pairing code
      parameters:
      - description: Instance data
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_instance_service.PairStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Pairing code
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Request pairing code
      tags:
      - Instance
  /instance/proxy/{instanceId}:
    delete:
      consumes:
      - application/json
      description: Delete proxy
      parameters:
      - description: Instance id
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Proxy deleted successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete proxy
      tags:
      - Instance
  /instance/qr:
    get:
      consumes:
      - application/json
      description: Get instance QR code
      produces:
      - application/json
      responses:
        "200":
          description: Instance QR code
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get instance QR code
      tags:
      - Instance
  /instance/status:
    get:
      consumes:
      - application/json
      description: Get instance status
      produces:
      - application/json
      responses:
        "200":
          description: Instance status
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get instance status
      tags:
      - Instance
  /label/chat:
    post:
      consumes:
      - application/json
      description: Add label to chat
      parameters:
      - description: Label data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_Zapbox-API_evolution-go_pkg_label_service.ChatLabelStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Add label to chat
      tags:
      - Label
...
  /newsletter/list:
    get:
      consumes:
      - application/json
      description: List newsletters
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List newsletters
      tags:
      - Newsletter
...
swagger: "2.0"